Add some ValidationSummary events tests (#11648)
